+++ restfy/request.py
@@ -2,11 +2,37 @@ import json
 from restfy.file import File
 
 
+class AccessControl:
+    allow_origin = '*'
+    expose_headers = []
+    max_age = 0.0
+    allow_credentials = False
+    allow_methods = []
+    allow_headers = []
+
+    def get_response_headers(self):
+        headers = {
+            'Access-Control-Allow-Origin': self.allow_origin,
+        }
+        if self.allow_methods:
+            headers['Access-Control-Allow-Methods'] = ','.join(self.allow_methods)
+        if self.allow_headers:
+            headers['Access-Control-Allow-Headers'] = ','.join(self.allow_headers)
+        if self.allow_credentials:
+            headers['Access-Control-Allow-Credentials'] = self.allow_credentials
+        if self.max_age:
+            headers['Access-Control-Max-Age'] = self.max_age
+        if self.expose_headers:
+            headers['Access-Control-Expose-Headers'] = self.expose_headers
+        return headers
+
+
 class Request:
     def __init__(self, method: str = 'GET', version: str = '1.1'):
         self.app = None
         self.method = method
         self.url = ''
+        self.port = ''
         self.version = version
         self.body = None
         self.type = ''
@@ -21,7 +47,8 @@ class Request:
         self.multipart = False
         self.boundary = ''
         self.data = {}
-        self.query_args = None
+        self.query_args: dict = {}
+        self.path_args: dict = {}
 
     def add_header(self, key, value):
         self.headers[key] = value